+++ src/main/java/com/technova/campussphereapi/repository/EventRepository.java
@@ -13,7 +13,7 @@ public interface EventRepository extends JpaRepository<Event, Integer> {
     Optional<Event> findByNameAndDescription(String name, String description);
 
     @Query(value = "SELECT * FROM fn_filer_events(?, ?, ?, ?)", nativeQuery = true)
-    List<Object[]> getEventsFiltered(BigDecimal precioMin, BigDecimal precioMax, String categoriaName, String ubicacion);
+    List<Object[]> getEventsFiltered(Float precioMin, Float precioMax, String categoriaName, String ubicacion);
   
     List<Event> findTop8ByOrderByCreatedAtDesc();
 }
