+++ src/main/java/com/technova/campussphereapi/repository/PriceRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 
 public interface PriceRepository extends JpaRepository<Price, Integer> {
 
-    Optional<Price> findByPriceAndDescription(BigDecimal price, String description);
+    Optional<Price> findByPriceAndDescription(Float price, String description);
 }
